requirements.txt
absl-py==2.1.0
annotated-types==0.6.0
anyio==4.3.0
asttokens==2.4.1
astunparse==1.6.3
attacut==1.0.6
attrs==23.2.0
backcall==0.2.0
backports.tarfile==1.0.0
bcrypt==4.1.2
beautifulsoup4==4.12.3
bleach==6.1.0
cachetools==5.3.3
certifi==2024.2.2
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cssselect==1.2.0
cssutils==2.10.2
decorator==5.1.1
defusedxml==0.7.1
docopt==0.6.2
docx2pdf==0.1.8
ecdsa==0.19.0
executing==2.0.1
fastapi==0.110.1
fastjsonschema==2.20.0
filelock==3.14.0
fire==0.6.0
flatbuffers==24.3.25
fsspec==2024.5.0
gast==0.5.4
google-api-core==2.18.0
google-api-python-client==2.125.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
google-pasta==0.2.0
googleapis-common-protos==1.63.0
greenlet==3.0.3
grpcio==1.64.0
h11==0.14.0
h5py==3.11.0
httplib2==0.22.0
huggingface-hub==0.23.2
idna==3.7
importlib_metadata==7.1.0
intel-openmp==2021.4.0
ipython==8.12.3
jaraco.classes==3.4.0
jaraco.context==5.3.0
jaraco.functools==4.0.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
jose==1.0.0
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
keras==3.3.3
keyring==25.1.0
libclang==18.1.1
lxml==5.2.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
mkl==2021.4.0
ml-dtypes==0.3.2
more-itertools==10.2.0
mpmath==1.3.0
mysqlclient==2.2.4
namex==0.0.8
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
networkx==3.3
nptyping==2.5.0
numpy==1.26.4
oauthlib==3.2.2
opt-einsum==3.3.0
optree==0.11.0
packaging==24.0
pandas==2.2.2
pandocfilters==1.5.1
parso==0.8.4
passlib==1.7.4
pickleshare==0.7.5
pillow==10.3.0
pip-check==2.9
pipreqs==0.5.0
platformdirs==4.2.2
premailer==3.10.0
prompt_toolkit==3.0.47
proto-plus==1.23.0
protobuf==4.25.3
pure-eval==0.2.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.7.0
pydantic_core==2.18.1
Pygments==2.18.0
PyMuPDF==1.24.9
PyMuPDFb==1.24.9
PyMySQL==1.1.0
pypandoc==1.13
pyparsing==3.1.2
pypiwin32==223
pythainlp==5.0.2
python-crfsuite==0.9.10
python-dateutil==2.9.0.post0
python-docx==1.1.0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
pytz==2024.1
pywin32==306
pywin32-ctypes==0.2.2
PyYAML==6.0.1
pyzmq==26.0.3
referencing==0.35.1
regex==2024.5.15
reportlab==4.2.0
requests==2.31.0
requests-oauthlib==2.0.0
rich==13.7.1
rpds-py==0.18.1
rsa==4.9
safetensors==0.4.3
scikit-learn==1.5.0
scipy==1.13.1
secure-smtplib==0.1.1
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.29
ssg==0.0.8
stack-data==0.6.3
starlette==0.37.2
sympy==1.12.1
tbb==2021.12.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-intel==2.16.1
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
terminaltables==3.1.10
threadpoolctl==3.5.0
tinycss2==1.3.0
tokenizers==0.19.1
torch==2.3.0
tornado==6.4.1
tqdm==4.66.2
traitlets==5.14.3
typing_extensions==4.11.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
uvicorn==0.29.0
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==3.0.3
wrapt==1.16.0
yagmail==0.15.293
yarg==0.1.9
zipp==3.18.1
